refactor: moved test connection to the common package
