+++ frontend/services/testconnectionotel/exporttester.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/odigos-io/odigos/destinations/testconnection"
+	"github.com/odigos-io/odigos/common/config/testconnection"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
